Fixed checkstyle and unittests
