+++ api/src/main/java/io/strimzi/api/kafka/model/authentication/KafkaClientAuthenticationScramSha512.java
@@ -27,10 +27,33 @@ public class KafkaClientAuthenticationScramSha512 extends KafkaClientAuthenticat
 
     public static final String TYPE_SCRAM_SHA_512 = "scram-sha-512";
 
+    private String username;
+    private PasswordSecretSource passwordSecret;
+
     @Description("Must be `" + TYPE_SCRAM_SHA_512 + "`")
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @Override
     public String getType() {
         return TYPE_SCRAM_SHA_512;
+    }
+
+    @Description("Reference to the `Secret` which holds the password.")
+    @Override
+    public PasswordSecretSource getPasswordSecret() {
+        return passwordSecret;
+    }
+
+    public void setPasswordSecret(PasswordSecretSource passwordSecret) {
+        this.passwordSecret = passwordSecret;
+    }
+
+    @Description("Username used for the authentication.")
+    @Override
+    public String getUsername() {
+        return username;
+    }
+
+    public void setUsername(String username) {
+        this.username = username;
     }
 }
